+++ services/121-service/src/payments/payments.module.ts
@@ -17,8 +17,12 @@ import { SafaricomModule } from '@121-service/src/payments/fsp-integration/safar
 import { PaymentEventsModule } from '@121-service/src/payments/payment-events/payment-events.module';
 import { PaymentsController } from '@121-service/src/payments/payments.controller';
 import { RedisModule } from '@121-service/src/payments/redis/redis.module';
-import { PaymentsHelperService } from '@121-service/src/payments/services/payments.helper.service';
-import { PaymentsService } from '@121-service/src/payments/services/payments.service';
+import { PaymentsExcelFspService } from '@121-service/src/payments/services/payments-excel-fsp.service';
+import { PaymentsExecutionService } from '@121-service/src/payments/services/payments-execution.service';
+import { PaymentsProgressHelperService } from '@121-service/src/payments/services/payments-progress.helper.service';
+import { PaymentsReportingHelperService } from '@121-service/src/payments/services/payments-reporting.helper.service';
+import { PaymentsReportingService } from '@121-service/src/payments/services/payments-reporting.service';
+import { TransactionJobsCreationService } from '@121-service/src/payments/services/transaction-jobs-creation.service';
 import { TransactionEntity } from '@121-service/src/payments/transactions/transaction.entity';
 import { TransactionsModule } from '@121-service/src/payments/transactions/transactions.module';
 import { ProgramFspConfigurationsModule } from '@121-service/src/program-fsp-configurations/program-fsp-configurations.module';
@@ -69,14 +73,18 @@ import { createScopedRepositoryProvider } from '@121-service/src/utils/scope/cre
     PaymentEventsModule,
   ],
   providers: [
-    PaymentsService,
-    PaymentsHelperService,
+    PaymentsExecutionService,
+    PaymentsReportingService,
+    PaymentsReportingHelperService,
+    PaymentsProgressHelperService,
+    TransactionJobsCreationService,
+    PaymentsExcelFspService,
     LookupService,
     InclusionScoreService,
     AzureLogService,
     createScopedRepositoryProvider(RegistrationAttributeDataEntity),
   ],
   controllers: [PaymentsController],
-  exports: [PaymentsService],
+  exports: [PaymentsExecutionService],
 })
 export class PaymentsModule {}
