+++ services/121-service/src/payments/services/payments-reporting.service.spec.ts
@@ -2,8 +2,8 @@ import { TestBed } from '@automock/jest';
 import { Repository } from 'typeorm';
 
 import { PaymentEntity } from '@121-service/src/payments/entities/payment.entity';
-import { PaymentsHelperService } from '@121-service/src/payments/services/payments.helper.service';
-import { PaymentsService } from '@121-service/src/payments/services/payments.service';
+import { PaymentsReportingHelperService } from '@121-service/src/payments/services/payments-reporting.helper.service';
+import { PaymentsReportingService } from '@121-service/src/payments/services/payments-reporting.service';
 import { TransactionStatusEnum } from '@121-service/src/payments/transactions/enums/transaction-status.enum';
 import { TransactionScopedRepository } from '@121-service/src/payments/transactions/transaction.scoped.repository';
 import { ProgramEntity } from '@121-service/src/programs/program.entity';
@@ -40,17 +40,19 @@ const mockTransactions = [
   createMockTransaction('102', 200, TransactionStatusEnum.success),
 ];
 
-describe('PaymentsService - getTransactions', () => {
-  let service: PaymentsService;
+describe('PaymentsReportingService - getTransactions', () => {
+  let service: PaymentsReportingService;
   let transactionScopedRepository: TransactionScopedRepository;
   let programRepository: Repository<ProgramEntity>;
   let registrationPaginationService: RegistrationsPaginationService;
   let programRegistrationAttributeRepository: ProgramRegistrationAttributeRepository;
-  let paymentsHelperService: PaymentsHelperService;
+  let paymentsHelperService: PaymentsReportingHelperService;
   let paymentRepository: Repository<PaymentEntity>;
 
   beforeEach(async () => {
-    const { unit, unitRef } = TestBed.create(PaymentsService).compile();
+    const { unit, unitRef } = TestBed.create(
+      PaymentsReportingService,
+    ).compile();
 
     transactionScopedRepository = unitRef.get(TransactionScopedRepository);
     programRepository = unitRef.get('ProgramEntityRepository');
@@ -68,7 +70,7 @@ describe('PaymentsService - getTransactions', () => {
     programRegistrationAttributeRepository = unitRef.get(
       ProgramRegistrationAttributeRepository,
     );
-    paymentsHelperService = unitRef.get(PaymentsHelperService);
+    paymentsHelperService = unitRef.get(PaymentsReportingHelperService);
     paymentRepository = unitRef.get('PaymentEntityRepository');
 
     jest.spyOn<any, any>(service, 'getTransactions');
